requirements.txt
fastapi==0.109.1
uvicorn==0.24.0
mysql-connector-python==9.2.0
python-multipart==0.0.6
pydantic[email]==2.5.0
bcrypt==4.1.2
python-dotenv==1.0.0
httpx==0.25.2
python-jose[cryptography]==3.3.0
authlib==1.2.1